20241025 log

+++ responseBean.go
@@ -1,7 +1,7 @@
 package result
 
 type ResponseBean struct {
-	Code uint32 `json:"code"`
+	Code int64  `json:"code"`
 	Msg  string `json:"msg"`
 	Data any    `json:"data"`
 }
@@ -10,7 +10,7 @@ type NullJson struct {
 }
 
 type ResponseErrorBean struct {
-	Code uint32 `json:"code"`
+	Code int64  `json:"code"`
 	Msg  string `json:"msg"`
 }
 
@@ -22,6 +22,6 @@ func Success(data any) *ResponseBean {
 	}
 }
 
-func Error(errCode uint32, errMsg string) *ResponseErrorBean {
+func Error(errCode int64, errMsg string) *ResponseErrorBean {
 	return &ResponseErrorBean{errCode, errMsg}
 }
